Initial commit
DVD commentary programming...

The state of this project at this point is all boiler plate. The
result of creating a new public repo on github and using stack version 1.9.3
to create a scaffold of a project. I have an idea of where I'd like
to end up, but I don't know at all how to get there :joy: