Digging into some gitlib examples
...and thiking I might have to rethink using that package (module?
library?). So this commit represents adding the dependencies required to
get the "First Steps" from the gitlib README.md
(https://github.com/jwiegley/gitlib/tree/master/gitlib).

The compiler did a lot of the work here, the packages I needed to add to
the project's yaml files, though not documented from what I could tell,
we're called out by the compiler. Well, except gitlib-libgit2 that I
could find on Hackage.

Had to install a couple libs with apt along the way:
- text-icu-0.7.0.1 needed libicu-dev
- hlibgit2-0.18.0.16 needed libssl-dev

And after all of that, I'm thinking that my inexperience with
Haskell, paired with a lack of gitlib documentation (read 'examples')
is probably a good reason for me to go with a more from-scratch / naive
approach. Get back to the shallow end of the pool :sweat_smile:

Thinking I may just learn a bit about the IO() object and read some
files in a .git directory, or... even run git commands and grab the
output? I'm actually a _little_ concerned those things might be harder
than learning to use gitlib ha :shrug: